Refactor, removed model
